Creates modules
